+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### 1.34.0 - 2022/04/22
+
+### Added
+
+* Simple and small `tw-tasks-jobs-test-spring-boot-starter` module to reduce a bit boilerplate in services testing jobs.
+
+### Changed
+
+* Using `CooperativeStickyAssignor` when it is detected that `kafka-clients` is `3.+`.
+
 #### 1.33.1 - 2022/04/21
 
 ### Fixed